// src/schemas/women-employment.ts
import { z } from "zod";

export const womenEmploymentSchema = z.object({
  firstName: z.string().min(1, "First name is required"),
  lastName: z.string().min(1, "Last name is required"),
  employmentType: z.enum(["INDIVIDUAL", "GROUP"]),
  sector: z.string().min(1, "Sector is required"),
  year: z.string().min(4, "Year is required"),
  remark: z.string().optional(),
});

export type WomenEmploymentSchemaType = z.infer<typeof womenEmploymentSchema>;